+++ src/components/content/CCDA/ccda_viewer/components/Header/data/getGuardian.ts
@@ -1,0 +1,35 @@
+import { isEmpty } from "lodash";
+import xpath from "xpath";
+import { getHumanName } from "../../../helpers";
+import { LabelValueType } from "../../../types";
+
+export const getGuardian = (
+  patientRole: Document
+): LabelValueType[] | undefined => {
+  const guardians = xpath.select(
+    "*[name()='patient']/*[name()='guardian']",
+    patientRole
+  ) as Document[];
+
+  if (isEmpty(guardians)) return undefined;
+
+  return guardians.map((guardian) => {
+    const guardianPerson = xpath.select1(
+      "*[name()='guardianPerson']",
+      guardian
+    ) as Document;
+
+    const relationship = String(
+      xpath.select1("string(*[name()='code']/@displayName)", guardian)
+    );
+    const name = getHumanName(
+      xpath.select("*[name()='name']", guardianPerson) as Document[]
+    );
+
+    const withRelationship = relationship ? ` - ${relationship}` : "";
+    return {
+      label: "",
+      value: `${name}${withRelationship}`,
+    };
+  });
+};