+++ src/components/content/CCDA/ccda_viewer/components/Patient/data/getReligion.ts
@@ -1,0 +1,9 @@
+import xpath from "xpath";
+
+export const getReligion = (patient: Document): string =>
+  String(
+    xpath.select1(
+      "string(*[name()='religiousAffiliationCode']/@displayName)",
+      patient
+    )
+  );