+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # ctw-component-library
 
+## 0.34.0
+
+### Minor Changes
+
+- be72b6a: Add CCDA document viewer to conditions in condition history that have these binary documents.
+
+## 0.33.0
+
+### Minor Changes
+
+- 13c80ac: Add hover actions in conditons 2.0.
+
 ## 0.32.1
 
 ### Patch Changes